requirements.txt
Django
djangorestframework
django-localflavor
django_extensions
Werkzeug
mysqlclient
django-mysql
django-countries
django-registration-redux

# Heroku things.
gunicorn
dj-static
dj-database-url
psycopg2
